use proper flavors
