+++ app/fdroid/org/koreader/launcher/ApkUpdater.kt
@@ -1,0 +1,10 @@
+package org.koreader.launcher
+
+import android.app.Activity
+
+object ApkUpdater {
+    @SuppressWarnings("unused")
+    fun download(activity: Activity, url: String, name: String): Int {
+        return -1
+    }
+}
