+++ app/manifest/rocks.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+     <uses-permission android:name="android.permission.REQUEST_INSTALL_PACKAGES" />
+</manifest>
